// app/CasterWeb/src/api/types.ts
export interface ServerState {
  uid: string;
  online_time: number;
  update_time: number;
  login_mpt: string;
  alias_mpt: string;
  type: number;
  account: string;
  ip: string;
  port: number;
  online_seconds: number;
  tcp_delay: number;
  ecef_x: number;
  ecef_y: number;
  ecef_z: number;
  position_update_time: number;
}

export interface ClientState {
  uid: string;
  online_time: number;
  update_time: number;
  login_mpt: string;
  alias_mpt: string;
  type: number;
  account: string;
  ip: string;
  port: number;
  online_seconds: number;
  tcp_delay: number;
  ecef_x: number;
  ecef_y: number;
  ecef_z: number;
  position_update_time: number;
  quality: number;
  sat_num: number;
  diff: number;
  distance: number;
}
// ...
export interface CasterNode {
  uid: string;
  node_name: string;
  set_version: string;
  tag_version: string;
  run_platform: string;
  cpu_usage: number;
  mem_usage: number;
  queue_delay: number;
  sub_ping_delay: number;
  sub_tcp_delay: number;
  pub_ping_delay: number;
  pub_tcp_delay: number;
  send_total: number;
  send_speed: number;
  recv_total: number;
  recv_speed: number;
  connect_count: number;
  server_count: number;
  client_count: number;
  online_time: number;
  update_time: number;
}